requirements.txt
# Core LangChain dependencies
langchain>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0
langchain-community>=0.2.0

# PDF processing
PyMuPDF>=1.23.0

# ML and data processing
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.66.0

# Embeddings
sentence-transformers>=2.5.0
langchain-community[embeddings]

# Vector Stores
qdrant-client>=1.7.0


# Environment and config
python-dotenv>=1.0.0
pydantic>=2.5.0

# Optional but recommended
jsonschema>=4.21.0
tiktoken>=0.6.0